src/requirements.txt
Babel==2.7.0
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
ebaysdk==2.1.5
essential-generators==0.9.2
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-Table==0.5.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.2.0
pycparser==2.19
PyMySQL==0.9.3
pytz==2019.3
randomuser==1.5.1
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.16.0
WTForms==2.2.1
